fix: implement basic operand and fixing

+++ Calculator.java
@@ -1,10 +1,10 @@
 public class Calculator {
 
-    public int add(int a, int b) {
+    public double add(double a, double b) {
         return a + b;
     }
 
-    public int subtract(int a, int b) {
+    public double subtract(double a, double b) {
         return a - b;
     }
